dom/tests/mochitest/pointerlock/file_mouselocklost.html

<!DOCTYPE HTML>
<html>
  <!-- https://bugzilla.mozilla.org/show_bug.cgi?id=633602 -->
  <head>
    <title>Test for Bug 633602</title>
    <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js">
    </script>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <script type="application/javascript" src="mouselock_util.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  </head>
  <body onload="start();">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
      Mozilla Bug 633602
    </a>
    <div id="content">
      <div style="background-color:blue; width:500px; height:500px;
        float: left;" id="blueBox">
        <p style="color:White" id='info'>
          MouseLock Element
        </p>
      </div>
    </div>
    <pre id="test">
      <script type="text/javascript">
        /** Test for Bug 633602 **/
        /** This test is to verify that the mouselocklost event fired by
            the browser has the bubbles attribute set to true, and
            the target attribute is the correct element **/
        SimpleTest.waitForExplicitFinish();
        var pointer = navigator.mozPointer;
        var blueBox = document.getElementById("blueBox");
        var firstTest = true;
        
        document.addEventListener("mozfullscreenchange", function() {
          if (document.mozFullScreen && 
              document.mozFullScreenElement === blueBox) {
            // If we've locked on to the blueBox element, proceed with test
            pointer.lock(blueBox, function(){
              ok(pointer.isLocked, "Manual unlock request - The mouselocklost event");
              pointer.unlock();
            });
          }/* else {
            ok(pointer.isLocked, "Element failed to enter fullscreen");
            SimpleTest.finish();
          }*/
        }, false);
        
        blueBox.addEventListener("pointerlocklost", function(e) {
          // First test scenario: calling pointer.unlock() and testing the
          // mouselockevent attributes
          if (firstTest) {
            firstTest = false;
            ok(e.bubbles, "Manual unlock request - The mouselocklost event " +
              "should have it's bubbles attribute set to true");
            is(e.target.id, "blueBox", "target should be blueBox");
            // Pending no problems, begin second test
            pointer.lock(blueBox, function() {
              document.mozCancelFullScreen();
            });
          } else if (!firstTest) {
            // Second test scenario: calling document.mozCancelFullScreen() 
            // and testing the mouselockevent attributes
            ok(e.bubbles, "mozCancelFullScreen request - The mouselocklost " +
              "event should have it's bubbles attribute set to true");
            is(e.target.id, "blueBox", "target should be blueBox");
            SimpleTest.finish();
          }
        }, false);
        
        function start()
        {
          SimpleTest.waitForFocus(function() {
            blueBox.mozRequestFullScreen();
          });
        }
      </script>
    </pre>
  </body>
</html>
